Unix/Windows common support
* Makes the implementation contain both Unix and Windows implementation on
  Windows. Used implementation can be used using env variable
  ANSITerminal=UNIX/WINDOWS
* Fix 'erase Screen' on Windows
