--- src/ANSITerminal_win.mli
@@ -1,1 +1,0 @@
-ANSITerminal.mli
